requirements.txt
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.1
pyOpenSSL==23.2.0
torch==1.13.0
transformers==4.11.3
easyocr==1.6.2
numpy==1.21.0